Fixed bug with player stats error java.lang.Double cannot be cast to java.lang.Integer

+++ core/src/main/java/com/lielamar/armsrace/listeners/shop/OnShopClick.java
@@ -48,7 +48,7 @@ public class OnShopClick implements Listener {
 					if (item.getLevels().size() == 0) return;
 
 					SkillLevel nextPlayerLevel = null;
-					int playerLevel = cp.getSkillLevel(item.getSkillType());
+					double playerLevel = cp.getSkillLevel(item.getSkillType());
 
 					for (SkillLevel lvl : item.getLevels()) {
 						if (playerLevel + 1 == lvl.getLevel())
